backend/requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
diffusers==0.25.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.47.0
fsspec==2023.12.2
gradio==4.15.0
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.9.0.80
orjson==3.9.12
packaging==23.2
pandas==2.1.4
Pillow==9.3.0
psutil==5.9.7
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.28.1
rich==13.7.0
rpds-py==0.17.1
ruff==0.1.13
safetensors==0.4.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.35.1
sympy==1.12
threadpoolctl==3.3.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.13
uvicorn==0.26.0
websockets==11.0.3
zipp==3.17.0
